Move coverage calculations into own module, implement station info
