+++ openlineplanner-rs/src/overlay.rs
@@ -15,6 +15,8 @@ use geojson::ser::to_feature_collection_string;
 use serde::Deserialize;
 use serde::Serialize;
 
+use crate::coverage::CoverageMap;
+
 #[derive(Deserialize)]
 pub enum OverlayName {
     Residence,
@@ -44,6 +46,54 @@ impl House {
     pub fn haversine_distance(&self, rhs: &Point) -> f64 {
         self.geometry.haversine_distance(rhs)
     }
+}
+
+#[derive(Serialize)]
+pub struct HouseCoverageCollection(Vec<HouseCoverage>);
+
+impl From<CoverageMap<'_, '_>> for HouseCoverageCollection {
+    fn from(value: CoverageMap<'_, '_>) -> Self {
+        Self(
+            value
+                .0
+                .into_iter()
+                .flat_map(|(station, sci)| {
+                    sci.houses.into_iter().map(|hi| HouseCoverage {
+                        geometry: hi.house.geometry,
+                        data_layer: "dl".to_owned(),
+                        distance: hi.distance,
+                        closest_station: station.to_owned(),
+                    })
+                })
+                .collect(),
+        )
+    }
+}
+
+impl Responder for HouseCoverageCollection {
+    type Body = BoxBody;
+
+    fn respond_to(self, _req: &actix_web::HttpRequest) -> actix_web::HttpResponse<Self::Body> {
+        match to_feature_collection_string(&self.0) {
+            Ok(body) => HttpResponse::Ok()
+                .content_type(ContentType::json())
+                .body(body),
+            Err(error) => HttpResponse::InternalServerError()
+                .body(format!("failed to get coverage collection: {}", error)),
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct HouseCoverage {
+    #[serde(
+        serialize_with = "serialize_geometry",
+        deserialize_with = "deserialize_geometry"
+    )]
+    geometry: Point,
+    data_layer: String,
+    distance: f64,
+    closest_station: String,
 }
 
 impl Overlays {
